Decouple API logic from networking
Extracts the actual network implementation and makes the service focus on
request serialization and response parsing.

Tests don't require HTTP stubbing anymore, and are synchronous.

It provides a default Alamofire network implementation, but it could easily be
integrated with any other network layer.

As this is just a proof of concept, it doesn't do error handling very well, and
I put all the extension code inside SiteService.